Drop chars in TextOutputDev with no unicode CMap
If the font has no unicode cmap and the text is encoded with Identity-H,
it's not possible to output text with the information we have, so rather than
potentially corrupting the textual output stream, the characters are dropped.

It may be possible to keep the characters if they happen to lie in the
printable spectrum, but my first priority is to fix crashes where the
glib API returns an inconsistent number of glyphs via
poppler_page_get_text and poppler_page_get_text_layout.