Drop chars in TextOutputDev with no unicode CMap
If the font has no unicode cmap, it's not possible to output text for
that encoding, so rather than potentially corrupting the textual output
stream, the characters are dropped.

It may be possible to keep the characters if they happen to lie in the
printable spectrum, but my first priority is to fix crashes where the
glib API returns an inconsistent number of glyphs via
poppler_page_get_text and poppler_page_get_text_layout.